pyproject.toml
[tool.poetry]
name = "django-aiogram-template"
version = "0.1.0"
description = ""
authors = ["Sherif Ragimov <sh.ragimov@idaproject.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
gunicorn = "^23.0.0"
aiogram = "^3.20.0.post0"
whitenoise = "^6.11.0"
redis = "^6.4.0"
celery = "^5.5.3"
psycopg2-binary = "^2.9.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
